Refactor debug_utils into after_aot and after_dynamo modules
There are no code changes but I did take the opportunity to
reorder and group the functions once they were placed in their
respective modules.


[ghstack-poisoned]
