+++ torch/_dynamo/debug_utils.py
@@ -3,20 +3,16 @@ import functools
 import getpass
 import logging
 import os
-import shutil
 import subprocess
+import tempfile
 import textwrap
-import uuid
 from collections import Counter
 from importlib import import_module
-from tempfile import TemporaryFile
 
 import torch
-import torch.fx as fx
 from torch._prims_common import is_float_dtype
 
 from . import config
-from .backends.registry import lookup_backend, register_debug_backend
 from .utils import clone_inputs, get_debug_dir
 
 log = logging.getLogger(__name__)
@@ -305,6 +301,7 @@ from torch._inductor.compile_fx import compile_fx_inner
 from torch._dynamo.debug_utils import same_two_models
 """
 
+
 COMPILER_REPRO_OPTIONS = {
     "inductor": (INDUCTOR_IMPORT, "compile_fx_inner", "inductor_fails"),
     "inductor_accuracy": (
@@ -313,123 +310,6 @@ COMPILER_REPRO_OPTIONS = {
         "inductor_accuracy_fails",
     ),
 }
-
-
-def dump_compiler_graph_state(gm, args, compiler_name):
-    subdir = os.path.join(minifier_dir(), "checkpoints")
-    if not os.path.exists(subdir):
-        os.makedirs(subdir, exist_ok=True)
-    file_name = os.path.join(subdir, f"{len(gm.graph.nodes)}.py")
-    log.warning(
-        "Writing checkpoint with %s nodes to %s", len(gm.graph.nodes), file_name
-    )
-    with open(file_name, "w") as fd:
-        save_graph_repro(fd, gm, args, compiler_name)
-    curdir = os.getcwd()
-    repro_path = os.path.join(curdir, "repro.py")
-    try:
-        shutil.copyfile(file_name, repro_path)
-        log.warning("Copying repro file for convenience to %s", repro_path)
-        if use_buck:
-            BuckTargetWriter(file_name).write()
-    except OSError:
-        log.warning("No write permissions for %s", repro_path)
-        pass
-
-
-def save_graph_repro(fd, gm, args, compiler_name):
-    sync_line = ""
-    for arg in args:
-        if isinstance(arg, torch.Tensor) and arg.is_cuda:
-            sync_line = "torch.cuda.synchronize() # Ensures that segfaults are surfaced"
-            break
-
-    if "inductor" in compiler_name:
-        fd.write("import torch._inductor.overrides\n")
-    fd.write(generate_compiler_repro_string(gm, args))
-    fd.write(COMPILER_REPRO_OPTIONS[compiler_name][0])
-    if "_accuracy" in compiler_name:
-        fd.write(
-            textwrap.dedent(
-                f"""
-                compiled = {COMPILER_REPRO_OPTIONS[compiler_name][1]}(mod, args)
-                class AccuracyError(Exception):
-                    pass
-                if not same_two_models(mod, compiled, args, only_fwd=True):
-                    raise AccuracyError("Bad accuracy detected")
-                """
-            )
-        )
-    else:
-        fd.write(
-            textwrap.dedent(
-                f"""
-                compiled = {COMPILER_REPRO_OPTIONS[compiler_name][1]}(mod, args)
-                ref = compiled(args)
-                {sync_line}
-                """
-            )
-        )
-
-
-def isolate_fails(fx_g, args, compiler_name: str, env=None, patch_code=None):
-    if env is None:
-        env = {}
-    subdir = os.path.join(os.getcwd(), "isolate")
-    if not os.path.exists(subdir):
-        os.makedirs(subdir, exist_ok=True)
-    file_name = os.path.join(subdir, f"{str(uuid.uuid4())[:5]}.py")
-    with open(file_name, "w") as fd:
-        repro_code = generate_compiler_repro_string(fx_g, args)
-        if patch_code is not None:
-            repro_code = repro_code.replace(TEST_REPLACEABLE_COMMENT, patch_code)
-        fd.write(repro_code)
-        fail_fn = COMPILER_REPRO_OPTIONS[compiler_name][2]
-        fd.write(
-            textwrap.dedent(
-                f"""
-                from {__name__} import {fail_fn}
-                """
-            )
-        )
-        fd.write(
-            textwrap.dedent(
-                f"""
-                if {fail_fn}(mod, args):
-                    exit(1)
-                else:
-                    exit(0)
-                """
-            )
-        )
-    # with open(file_name, "r") as fd:
-    #     print(fd.read())
-    new_env = os.environ.copy()
-    new_env = {**new_env, **env}
-    stdout, stderr = TemporaryFile(), TemporaryFile()
-
-    if use_buck:
-        cmd = BuckTargetWriter(file_name).write(print_msg=False)
-    else:
-        cmd = ["python", file_name]
-
-    p = subprocess.Popen(
-        cmd,
-        cwd=subdir,
-        stdout=stdout,
-        stderr=stderr,
-        env=new_env,
-    )
-    p.wait()
-
-    if p.returncode != 0:
-        stdout.seek(0)
-        stderr.seek(0)
-        print(textwrap.indent(stdout.read().decode("utf-8"), prefix=">>  "))
-        print(textwrap.indent(stderr.read().decode("utf-8"), prefix=">>  "))
-        # print(f"Isolated test failed - {file_name}")
-        return True
-    return False
 
 
 def inductor_fails(fx_g, args, check_str=None):
@@ -467,12 +347,6 @@ def inductor_fails(fx_g, args, check_str=None):
     return False
 
 
-def inductor_accuracy_fails(fx_g, args, check_str=None):
-    from torch._inductor.compile_fx import compile_fx_inner
-
-    return backend_aot_accuracy_fails(fx_g, args, compile_fx_inner)
-
-
 def get_minifier_repro_path():
     return os.path.join(minifier_dir(), "minifier_launcher.py")
 
@@ -492,146 +366,8 @@ def helper_for_dump_minify(contents):
         raise NotImplementedError("Could not write to {minified_repro_path}") from e
 
 
-def dump_to_minify(gm, args, compiler_name: str):
-    favored_device = 1 if torch.cuda.device_count() >= 2 else 0
-
-    contents = textwrap.dedent(
-        f"""
-isolate_fails_code_str = None
-
-{generate_compiler_repro_string(gm, args)}
-
-from functools import partial
-from {__name__} import (
-    isolate_fails,
-    dump_compiler_graph_state,
-)
-from functorch.compile import minifier
-
-env_variables = {{"CUDA_VISIBLE_DEVICES": "{favored_device}"}}
-
-minifier(
-    mod,
-    args,
-    module_fails=partial(isolate_fails, env=env_variables, compiler_name="{compiler_name}", patch_code=isolate_fails_code_str),
-    dump_state=partial(dump_compiler_graph_state, compiler_name="{compiler_name}"),
-)
-        """
-    )
-    return helper_for_dump_minify(contents)
-
-
 class AccuracyError(Exception):
     pass
-
-
-def wrap_compiler_debug(unconfigured_compiler_fn, compiler_name: str):
-    """
-    Minifier for Fx Graph modules after Aot Autograd has finished. We wrap both
-    forward and backward call separately with the backend compiler_fn - like
-    inductor or nvfuser. Intercepting after Aot Autograd presents neat
-    abstraction, where all the params are lifted as graph inputs, making it easy
-    to save the graph as a string.
-    """
-
-    @functools.wraps(unconfigured_compiler_fn)
-    def debug_wrapper(gm, example_inputs, **kwargs):
-        from torch._subclasses import FakeTensorMode
-
-        compiler_fn = functools.partial(unconfigured_compiler_fn, **kwargs)
-
-        orig_graph = copy.deepcopy(gm.graph)
-        assert config.repro_after in ("dynamo", "aot", None)
-        inner_compiled_fn = None
-
-        def deferred_for_real_inputs(real_inputs):
-            """
-            Aot Autograd fw_compiler and bw_compiler can have fake tensors. So,
-            example_inputs can be fake tensors. We can call compiler_fn (which is
-            inductor or nvfuser) with fake tensors but the actually compiled_fn
-            should be called with real tensors. Therefore, the actual invocation
-            is deferred.
-            """
-            # Avoid re-compiling when we call the compiled function twice. This happens
-            # when we run the model inference or training in a for loop like here
-            # https://github.com/pytorch/torchdynamo/issues/1687#issuecomment-1280040633
-            nonlocal inner_compiled_fn
-            # Copy the tensor attrs like shape, stride etc by converting to Fake Tensor
-            # because inductor clears the tensor list in its codegen. And example_inputs
-            # are available only for the first invocation.
-            fake_mode = FakeTensorMode()
-            copy_tensor_attrs = [
-                fake_mode.from_tensor(x) if isinstance(x, torch.Tensor) else x
-                for x in real_inputs
-            ]
-            if config.repro_level == 3:
-                # Always dump the original module in case we have segfaults
-                dump_to_minify(
-                    fx.GraphModule(gm, orig_graph), real_inputs, compiler_name
-                )
-
-            if config.repro_level == 4:
-                if compiler_name != "inductor":
-                    raise NotImplementedError(
-                        "Accuracy minification is supported for inductor only"
-                    )
-                if inner_compiled_fn is None:
-                    inner_compiled_fn = compiler_fn(gm, example_inputs)
-                if backend_aot_accuracy_fails(gm, real_inputs, compiler_fn):
-                    log.warning("Accuracy failed for the AOT Autograd graph")
-                    dump_compiler_graph_state(
-                        fx.GraphModule(gm, orig_graph),
-                        copy_tensor_attrs,
-                        f"{compiler_name}_accuracy",
-                    )
-                    dump_to_minify(
-                        fx.GraphModule(gm, orig_graph),
-                        copy_tensor_attrs,
-                        f"{compiler_name}_accuracy",
-                    )
-                    raise AccuracyError("Bad accuracy detected")
-                else:
-                    # Call the compiled function with real inputs
-                    return inner_compiled_fn(real_inputs)
-            else:
-                try:
-                    # Call the compiler_fn - which is either aot_autograd or inductor
-                    # with fake inputs
-                    if inner_compiled_fn is None:
-                        inner_compiled_fn = compiler_fn(gm, example_inputs)
-                    # Call the compiled function with real inputs
-                    out = inner_compiled_fn(real_inputs)
-                    # sync cuda kernels to ensure IMA detection
-                    for arg in example_inputs:
-                        if isinstance(arg, torch.Tensor) and arg.is_cuda:
-                            torch.cuda.synchronize()
-                            break
-                    return out
-                except Exception as e:
-                    if config.repro_level == 1:
-                        dump_compiler_graph_state(
-                            fx.GraphModule(gm, orig_graph),
-                            copy_tensor_attrs,
-                            compiler_name,
-                        )
-                    elif config.repro_level == 2:
-                        dump_to_minify(
-                            fx.GraphModule(gm, orig_graph),
-                            copy_tensor_attrs,
-                            compiler_name,
-                        )
-                    log.error("CompilerError")
-                    raise
-
-        if config.repro_after == "aot":
-            compiled_fn = deferred_for_real_inputs
-            compiled_fn._boxed_call = True
-        else:
-            compiled_fn = compiler_fn(gm, example_inputs)
-
-        return compiled_fn
-
-    return debug_wrapper
 
 
 def run_fwd_maybe_bwd(gm, args, only_fwd=False):
@@ -759,163 +495,6 @@ def cast_to_fp64(model, inputs):
     return cast_to(torch.float64, model, inputs)
 
 
-def generate_dynamo_fx_repro_string(
-    model_str, args, compiler_name, check_accuracy=False
-):
-    """
-    Generate a repro string for backend-agnostic minified version.
-    """
-
-    run_code = textwrap.dedent(
-        f"""
-with torch.cuda.amp.autocast(enabled={torch.is_autocast_enabled()}):
-    ref = run_fwd_maybe_bwd(mod, args)
-    res = run_fwd_maybe_bwd(opt_mod, args)
-    """
-    )
-
-    if config.repro_level == 4 or check_accuracy:
-        run_code = textwrap.dedent(
-            f"""
-mod.eval()
-opt_mod.eval()
-
-class AccuracyError(Exception):
-    pass
-
-with torch.cuda.amp.autocast(enabled={torch.is_autocast_enabled()}):
-    assert same_two_models(mod, mod, args), "Eager itself failed"
-    if not same_two_models(mod, opt_mod, args):
-        raise AccuracyError("Dynamo failed")
-    """
-        )
-
-    return textwrap.dedent(
-        f"""
-from math import inf
-import torch
-from torch import tensor, device
-import torch.fx as fx
-import torch._dynamo
-from torch._dynamo.testing import rand_strided
-from torch._dynamo.debug_utils import run_fwd_maybe_bwd
-from torch._dynamo.debug_utils import same_two_models
-
-{generate_config_string()}
-
-{TEST_REPLACEABLE_COMMENT}
-{extra_imports}
-
-args = {[(tuple(a.shape), tuple(a.stride()), a.dtype, a.device.type, a.requires_grad) for a in args]}
-args = [rand_strided(sh, st, dt, dev).requires_grad_(rg) for (sh, st, dt, dev, rg) in args]
-
-{model_str}
-
-mod = Repro()
-opt_mod = torch._dynamo.optimize("{compiler_name}")(mod)
-
-{run_code}
-        """
-    )
-
-
-def dump_backend_repro_as_file(gm, args, compiler_name, check_accuracy=False):
-    """
-    Saves the repro to a repro.py file
-    """
-    curdir = os.getcwd()
-    subdir = os.path.join(os.getcwd(), "checkpoints")
-    if not os.path.exists(subdir):
-        os.makedirs(subdir, exist_ok=True)
-    file_name = os.path.join(subdir, f"minified_{len(gm.graph.nodes)}_nodes.py")
-    log.warning(
-        "Writing checkpoint with %s nodes to %s", len(gm.graph.nodes), file_name
-    )
-
-    model_str = NNModuleToString.convert(gm)
-    with open(file_name, "w") as fd:
-        fd.write(
-            generate_dynamo_fx_repro_string(
-                model_str, args, compiler_name, check_accuracy
-            )
-        )
-    latest_repro = os.path.join(curdir, "repro.py")
-    log.warning("Copying %s to %s for convenience", file_name, latest_repro)
-
-    if use_buck:
-        BuckTargetWriter(latest_repro).write()
-
-    shutil.copyfile(file_name, latest_repro)
-
-
-# TODO - Commented because we are assuming that nn.Modules can be safely repr'd
-# If that does not work, we might have to bring this code back. So, keeping it
-# as it is for now.
-# def dump_backend_repro_as_tarfile(gm, args, compiler_name):
-#     """
-#     Saves the repro in repro.tar.gz, as opposed to a file. This is used for
-#     cases, where we can't convert a Fx GraphModule to a string, and therefore
-#     fallback to to_folder for serialization. We accompany this with a repro.py
-#     script that imports the saved module, sets it up and runs the model to repro
-#     the error.
-#     """
-#     import tarfile
-
-#     subdir = os.path.join(minifier_dir(), "checkpoints")
-#     if not os.path.exists(subdir):
-#         os.makedirs(subdir, exist_ok=True)
-
-#     tmp_dir = os.path.join(subdir, f"{len(gm.graph.nodes)}")
-#     if os.path.exists(tmp_dir):
-#         shutil.rmtree(tmp_dir)
-#     os.makedirs(tmp_dir, exist_ok=True)
-
-#     file_name = os.path.join(tmp_dir, "repro.py")
-#     gm_dir = os.path.join(tmp_dir, "module")
-#     if not os.path.exists(gm_dir):
-#         os.makedirs(gm_dir, exist_ok=True)
-#     for node in gm.graph.nodes:
-#         new_kwargs = {}
-#         for k, v in node.kwargs.items():
-#             if isinstance(v, torch.device):
-#                 v = v.type
-#             new_kwargs[k] = v
-#         node.kwargs = new_kwargs
-#     gm.recompile()
-
-#     print(f"Writing checkpoint with {len(gm.graph.nodes)} nodes to {file_name}")
-#     with open(file_name, "w") as fd:
-#         # TODO - Add the readable version of to_folder when available
-#         gm.to_folder(gm_dir, "Repro")
-#         fd.write(
-#             generate_dynamo_fx_repro_string(
-#                 "from module import Repro", args, compiler_name
-#             )
-#         )
-
-#     local_dir = os.path.join(config.base_dir, "repro")
-#     if os.path.exists(local_dir):
-#         shutil.rmtree(local_dir)
-#     shutil.copytree(tmp_dir, local_dir)
-#     local_tar_file = os.path.join(config.base_dir, "repro.tar.gz")
-#     print(f"Writing checkpoint with {len(gm.graph.nodes)} locally to {local_tar_file}")
-#     with tarfile.open(local_tar_file, "w:gz") as tar:
-#         tar.add(local_dir, arcname=os.path.basename(local_dir))
-
-
-def dump_backend_state(gm, args, compiler_name, check_accuracy=False):
-    """
-    Dumps the dynamo graph to repro the issue.
-    1) It tries to convert Fx GraphModule to a string. If we can, it writes to a
-    repro.py file.
-    2) If we can't convert Fx GraphModule to a string, we use to_folder to save
-    the module and save a tar file.
-    """
-    assert NNModuleToString.can_convert_to_string(gm)
-    return dump_backend_repro_as_file(gm, args, compiler_name, check_accuracy)
-    # return dump_backend_repro_as_tarfile(gm, args, compiler_name)
-
-
 def backend_accuracy_fails(gm, example_inputs, compiler_fn, only_fwd=False):
     try:
         compiled_gm = compiler_fn(copy.deepcopy(gm), clone_inputs(example_inputs))
@@ -932,237 +511,3 @@ def backend_accuracy_fails(gm, example_inputs, compiler_fn, only_fwd=False):
         return False
 
     return not same_two_models(gm, compiled_gm, example_inputs, only_fwd)
-
-
-backend_aot_accuracy_fails = functools.partial(backend_accuracy_fails, only_fwd=True)
-
-
-def backend_fails(gm, example_inputs, compiler_fn, orig_failure):
-    """
-    Minifier uses this function to identify if the minified graph module fails
-    with the same error.
-
-    One caveat is that minifier can potentially go into a wrong direction when
-    the resulting graph module fails for a different reason. To avoid this, we
-    save the string for the original exception and check similarity between new
-    and old exception. They can be somewhat different in some cases, when the
-    exception string depends on the failing node information. So, we have a
-    loose similarity metric to guide the minifier path.
-    """
-    from difflib import SequenceMatcher
-
-    try:
-        compiled_gm = compiler_fn(gm, example_inputs)
-        run_fwd_maybe_bwd(compiled_gm, clone_inputs(example_inputs))
-        return False
-    except Exception as e:
-        new_failure = str(e)
-        if SequenceMatcher(None, orig_failure, new_failure).ratio() > 0.5:
-            return True
-        return False
-
-
-def dump_to_minify_after_dynamo(gm, args, compiler_name):
-    model_str = NNModuleToString.convert(gm)
-
-    minifier_backend = "dynamo_minifier_backend"
-    if config.repro_level == 4:
-        minifier_backend = "dynamo_accuracy_minifier_backend"
-
-    custom_compiler_error = (
-        textwrap.dedent(
-            """\
-        raise RuntimeError(
-            'Compiler name is None - this likely means that a custom compiler '
-            'was called by torchdynamo. Please remove this error, import your '
-            'custom compiler function, and replace the compiler_name="None" '
-            'line below to compiler_name=<my_imported_custom_function>'
-        )
-        """
-        )
-        if compiler_name is None
-        else ""
-    )
-
-    contents = textwrap.dedent(
-        f"""
-import os
-from math import inf
-import torch
-from torch import tensor, device
-import torch.fx as fx
-import functools
-import torch._dynamo
-from torch._dynamo.debug_utils import run_fwd_maybe_bwd
-from torch._dynamo.backends.registry import lookup_backend
-from torch._dynamo.testing import rand_strided
-
-{generate_config_string()}
-
-{TEST_REPLACEABLE_COMMENT}
-{extra_imports}
-
-args = {[(tuple(a.shape), tuple(a.stride()), a.dtype, a.device.type, a.requires_grad) for a in args]}
-args = [rand_strided(sh, st, dt, dev).requires_grad_(rg) for (sh, st, dt, dev, rg) in args]
-
-{model_str}
-mod = Repro()
-
-# Setup debug minifier compiler
-compiler_fn = lookup_backend("{minifier_backend}")
-{custom_compiler_error}
-dynamo_minifier_backend = functools.partial(
-    compiler_fn,
-    compiler_name="{compiler_name}",
-)
-opt_mod = torch._dynamo.optimize(dynamo_minifier_backend)(mod)
-
-with torch.cuda.amp.autocast(enabled={torch.is_autocast_enabled()}):
-    opt_mod(*args)
-        """
-    )
-    helper_for_dump_minify(contents)
-
-
-def wrap_backend_debug(unconfigured_compiler_fn, compiler_name: str):
-    """
-    A minifier decorator that wraps the TorchDynamo produced Fx graph modules.
-    As opposed to wrap_compiler_debug, this wrapper intercepts at the
-    TorchDynamo produced Fx Graph Module. This makes it backend-agnostic to some
-    level, e.g., it is useful for minifying issues related to Aot Autograd
-    tracing.  If an error is found, we minify and save the minified repro in
-    repro.tar.gz.
-    """
-
-    @functools.wraps(unconfigured_compiler_fn)
-    def debug_wrapper(gm, example_inputs, **kwargs):
-        compiler_fn = functools.partial(unconfigured_compiler_fn, **kwargs)
-        assert config.repro_after in ("dynamo", "aot", None)
-
-        if config.repro_after == "dynamo":
-
-            def add_paths(exc):
-                exc.minifier_path = os.path.join(minifier_dir(), "minifier_launcher.py")
-                if use_buck:
-                    exc.buck_command = " ".join(
-                        BUCK_CMD_PREFIX
-                        + [BuckTargetWriter(exc.minifier_path).cmd_line_path]
-                    )
-
-            if config.repro_level == 3:
-                dump_to_minify_after_dynamo(gm, example_inputs, compiler_name)
-
-            # Check for either accuracy (level 4) or other type of failures.
-            if config.repro_level == 4:
-                # Check Accuracy
-                compiled_gm = compiler_fn(copy.deepcopy(gm), example_inputs)
-                if backend_accuracy_fails(gm, example_inputs, compiler_fn):
-                    log.warning(
-                        "Accuracy failed for the TorchDynamo produced graph. Creating script to minify the error."
-                    )
-                    dump_to_minify_after_dynamo(
-                        fx.GraphModule(gm, copy.deepcopy(gm.graph)),
-                        example_inputs,
-                        compiler_name,
-                    )
-                    exc = AccuracyError("Bad accuracy detected.")
-                    add_paths(exc)
-                    raise exc
-            else:
-                try:
-                    compiled_gm = compiler_fn(copy.deepcopy(gm), example_inputs)
-                    run_fwd_maybe_bwd(compiled_gm, example_inputs)
-                except Exception as exc:
-                    log.warning(
-                        "Compiled Fx GraphModule failed. Creating script to minify the error."
-                    )
-                    if config.repro_level == 1:
-                        dump_state_fn = functools.partial(
-                            dump_backend_state, compiler_name=compiler_name
-                        )
-                        dump_state_fn(
-                            fx.GraphModule(gm, copy.deepcopy(gm.graph)), example_inputs
-                        )
-                    elif config.repro_level == 2:
-                        dump_to_minify_after_dynamo(
-                            fx.GraphModule(gm, copy.deepcopy(gm.graph)),
-                            example_inputs,
-                            compiler_name,
-                        )
-                    add_paths(exc)
-                    raise
-        else:
-            compiled_gm = compiler_fn(gm, example_inputs)
-
-        return compiled_gm
-
-    debug_wrapper._torchdynamo_orig_callable = unconfigured_compiler_fn
-
-    return debug_wrapper
-
-
-@register_debug_backend
-def dynamo_minifier_backend(gm, example_inputs, compiler_name):
-    from functorch.compile import minifier
-
-    compiler_fn = lookup_backend(compiler_name)
-
-    try:
-        compiled_gm = compiler_fn(gm, example_inputs)
-        run_fwd_maybe_bwd(compiled_gm, example_inputs)
-        raise ValueError("No issue was detected")
-    except Exception as exc:
-        orig_failure = str(exc)
-        log.warning(
-            "Compiled Fx GraphModule failed. Creating script to minify the error."
-        )
-        dump_state_fn = functools.partial(
-            dump_backend_state, compiler_name=compiler_name
-        )
-        dump_state_fn(fx.GraphModule(gm, copy.deepcopy(gm.graph)), example_inputs)
-        fails_fn = functools.partial(
-            backend_fails,
-            compiler_fn=compiler_fn,
-            orig_failure=orig_failure,
-        )
-        minifier(
-            gm,
-            example_inputs,
-            module_fails=fails_fn,
-            dump_state=dump_state_fn,
-        )
-    return gm
-
-
-@register_debug_backend
-def dynamo_accuracy_minifier_backend(gm, example_inputs, compiler_name):
-    from functorch.compile import minifier
-
-    compiler_fn = lookup_backend(compiler_name)
-
-    # Set the eval mode to remove randomness.
-    gm.eval()
-
-    # Check Accuracy
-    if backend_accuracy_fails(
-        gm, example_inputs, compiler_fn, only_fwd=config.repro_forward_only
-    ):
-        log.warning("Accuracy failed for the TorchDynamo produced graph")
-        dump_state_fn = functools.partial(
-            dump_backend_state, compiler_name=compiler_name, check_accuracy=True
-        )
-        fails_fn = functools.partial(
-            backend_accuracy_fails,
-            compiler_fn=compiler_fn,
-            only_fwd=config.repro_forward_only,
-        )
-        dump_state_fn(fx.GraphModule(gm, copy.deepcopy(gm.graph)), example_inputs)
-        minifier(
-            gm,
-            example_inputs,
-            module_fails=fails_fn,
-            dump_state=dump_state_fn,
-        )
-    else:
-        log.error("Input graph does not fail accuracy testing")
-    return gm
